Split guard clause tests into different files
